Added support for PDBBind data/mol2 files

+++ pdbqt_batch_converter.py
@@ -4,8 +4,9 @@ import subprocess
 from tqdm import tqdm
 
 MOAD_fatal_error_list = list()
+PDB_fatal_error_list = list()
 
-pdb_files_path = ''
+pdb_files_path = '/home/milesm/Dissertation/Data/Parsed/Non_redundant/PDBBind/'
 moad_files_path = '/home/milesm/Dissertation/Data/Parsed/Non_redundant/Binding_MOAD/'
 dude_files_path = ''
 
@@ -18,7 +19,7 @@ example_ligand_file = '/home/milesm/Desktop/4yki_ligand.pdb'
 example_protein_file = '/home/milesm/Desktop/7cpa/7cpa_protein.pdb'
 example_destination_path = '/home/milesm/Desktop/7cpa/'
 
-#pdb_structure_folders = [(pdb_files_path + folder) for folder in os.listdir(pdb_files_path)]
+pdb_structure_folders = [(pdb_files_path + folder) for folder in os.listdir(pdb_files_path)]
 moad_structure_folders = [(moad_files_path + folder) for folder in os.listdir(moad_files_path)]
 #dude_structure_folders = [(dude_files_path + folder) for folder in os.listdir(dude_files_path)]
 
@@ -46,41 +47,61 @@ def repair_stacked_ligand(ligand_filepath, ligand_file):
             print(f'Found pre-repaired ligand from a previous run: "{repair_check[0]}"')
         return repaired_ligand_filepath
 
-def autodock_convert(folder_name, destination_path, prep_ligand_command, prep_protein_command, ligand_filepath, ligand_file, receptor_filepath, receptor_file):
+def autodock_convert(folder_name, destination_path, prep_ligand_command, prep_protein_command, ligand_filepath, ligand_file, receptor_filepath, receptor_file, database_name):
     try:
         os.mkdir(f'{destination_path}{folder_name}')
     except FileExistsError:
         pass
     print(folder_name)
+    if database_name == 'pdb':
+        ligand_outfile = ligand_file.replace('mol2','pdb')
+    else:
+        ligand_outfile = ligand_file
     print('Preparing ligand...')
     try:
-        output = subprocess.check_output(f'{prep_ligand_command} -l {ligand_filepath} -A hydrogens -o {destination_path}{folder_name}/{ligand_file}qt', shell=True, stderr=subprocess.STDOUT)
+        output = subprocess.check_output(f'{prep_ligand_command} -l {ligand_filepath} -A hydrogens -o {destination_path}{folder_name}/{ligand_outfile}qt', shell=True, stderr=subprocess.STDOUT)
     except:
         try:
             print('REPAIRING LIGAND...')
             repaired_ligand_filepath = repair_stacked_ligand(ligand_filepath, ligand_file)
             if repaired_ligand_filepath is not None:
-                output = subprocess.check_output(f'{prep_ligand_command} -l {repaired_ligand_filepath} -A hydrogens -o {destination_path}{folder_name}/{ligand_file}qt', shell=True)
+                output = subprocess.check_output(f'{prep_ligand_command} -l {repaired_ligand_filepath} -A hydrogens -o {destination_path}{folder_name}/{ligand_outfile}qt', shell=True)
                 print('REPAIR SUCCESSFUL!')
         except:
-            MOAD_fatal_error_list.append(folder_name)
-            if os.path.isfile('MOAD_fatal_error_list.txt'):
-                os.remove('MOAD_fatal_error_list.txt')
-            with open('MOAD_fatal_error_list.txt','a+') as error_list:
-                error_list.write(str(MOAD_fatal_error_list))
-                error_list.close()
+            if database_name == 'moad':
+                MOAD_fatal_error_list.append(folder_name)
+                if os.path.isfile('MOAD_fatal_error_list.txt'):
+                    os.remove('MOAD_fatal_error_list.txt')
+                with open('MOAD_fatal_error_list.txt','a+') as error_list:
+                    error_list.write(str(MOAD_fatal_error_list))
+                    error_list.close()
+            if database_name == 'pdb':
+                PDB_fatal_error_list.append(folder_name)
+                if os.path.isfile('PDB_fatal_error_list.txt'):
+                    os.remove('PDB_fatal_error_list.txt')
+                with open('PDB_fatal_error_list.txt','a+') as error_list:
+                    error_list.write(str(PDB_fatal_error_list))
+                    error_list.close()
             print(f'FATAL PROBLEM WITH {folder_name}: Added to list and skipping...')
 
     print('Preparing protein...')
     try:
         output = subprocess.check_output(f'{prep_protein_command} -r {receptor_filepath} -A hydrogens -o {destination_path}{folder_name}/{receptor_file}qt -U waters', shell=True)
     except:
-        MOAD_fatal_error_list.append(folder_name)
-        if os.path.isfile('MOAD_fatal_error_list.txt'):
-            os.remove('MOAD_fatal_error_list.txt')
-        with open('MOAD_fatal_error_list.txt','a+') as error_list:
-            error_list.write(str(MOAD_fatal_error_list))
-            error_list.close()
+        if database_name == 'moad':
+            MOAD_fatal_error_list.append(folder_name)
+            if os.path.isfile('MOAD_fatal_error_list.txt'):
+                os.remove('MOAD_fatal_error_list.txt')
+            with open('MOAD_fatal_error_list.txt','a+') as error_list:
+                error_list.write(str(MOAD_fatal_error_list))
+                error_list.close()
+        if database_name == 'pdb':
+            PDB_fatal_error_list.append(folder_name)
+            if os.path.isfile('PDB_fatal_error_list.txt'):
+                os.remove('PDB_fatal_error_list.txt')
+            with open('PDB_fatal_error_list.txt','a+') as error_list:
+                error_list.write(str(PDB_fatal_error_list))
+                error_list.close()
         print(f'FATAL PROBLEM WITH {folder_name}: Added to list and skipping...')
 
 def batch_convert_to_pdbqt(structure_folders, database_name, destination_path):
@@ -99,7 +120,23 @@ def batch_convert_to_pdbqt(structure_folders, database_name, destination_path):
                 ligand_filepath = folder + '/' + ligand_file
                 receptor_file = [filename for filename in files_in_folder if 'protein.pdb' in filename and '.pdbqt' not in filename][0]
                 receptor_filepath = folder + '/' + receptor_file
-                autodock_convert(folder_name, destination_path, prep_ligand_command, prep_protein_command, ligand_filepath, ligand_file, receptor_filepath, receptor_file)
+                autodock_convert(folder_name, destination_path, prep_ligand_command, prep_protein_command, ligand_filepath, ligand_file, receptor_filepath, receptor_file, database_name)
                 pbar.update(1)
 
-batch_convert_to_pdbqt(moad_structure_folders[415:], 'moad', destination_path)
+    if database_name == 'pdb':
+        destination_path = destination_path + 'PDBBind/'
+        try:
+            os.mkdir(destination_path)
+        except FileExistsError:
+            pass
+        with tqdm(total=len(structure_folders)) as pbar:
+            for folder in structure_folders:
+                folder_name = folder.split('/')
+                folder_name = folder_name[len(folder_name) - 1]
+                files_in_folder = os.listdir(folder)
+                ligand_file = [filename for filename in files_in_folder if 'ligand.mol2' in filename and '.pdbqt' not in filename][0]
+                ligand_filepath = folder + '/' + ligand_file
+                receptor_file = [filename for filename in files_in_folder if 'protein.pdb' in filename and '.pdbqt' not in filename][0]
+                receptor_filepath = folder + '/' + receptor_file
+                autodock_convert(folder_name, destination_path, prep_ligand_command, prep_protein_command, ligand_filepath, ligand_file, receptor_filepath, receptor_file, database_name)
+                pbar.update(1)
